# src/main/java/com/example/expensetracker/summary/dto/CategoryTotalItem.java
package com.example.expensetracker.summary.dto;

import java.math.BigDecimal;

public record CategoryTotalItem(
        Long categoryId,
        String categoryName,
        BigDecimal total,
        BigDecimal percentage
) {
}